Factor out getSuggestionsFromOpenAI into a hook
