+++ projects/plugins/jetpack/extensions/blocks/ai-assistant/use-suggestions-from-openai.js
@@ -1,0 +1,148 @@
+import apiFetch from '@wordpress/api-fetch';
+import { useSelect } from '@wordpress/data';
+import { useEffect, useState } from '@wordpress/element';
+import { __ } from '@wordpress/i18n';
+import MarkdownIt from 'markdown-it';
+
+const useSuggestionsFromOpenAI = ( {
+	clientId,
+	content,
+	createPrompt,
+	setAttributes,
+	setErrorMessage,
+	tracks,
+	userPrompt,
+} ) => {
+	const [ isLoadingCategories, setIsLoadingCategories ] = useState( false );
+	const [ isLoadingCompletion, setIsLoadingCompletion ] = useState( false );
+	const [ showRetry, setShowRetry ] = useState( false );
+	// Let's grab post data so that we can do something smart.
+
+	const currentPostTitle = useSelect( select =>
+		select( 'core/editor' ).getEditedPostAttribute( 'title' )
+	);
+
+	//TODO: decide if we still want to load categories and tags now user is providing the prompt by default.
+	// If not the following can be removed.
+	let loading = false;
+	const categories =
+		useSelect( select => select( 'core/editor' ).getEditedPostAttribute( 'categories' ) ) || [];
+
+	const categoryObjects = useSelect(
+		select => {
+			return categories
+				.map( categoryId => {
+					const category = select( 'core' ).getEntityRecord( 'taxonomy', 'category', categoryId );
+
+					if ( ! category ) {
+						// Data is not yet loaded
+						loading = true;
+						return;
+					}
+
+					return category;
+				} )
+				.filter( Boolean ); // Remove undefined values
+		},
+		[ categories ]
+	);
+
+	const tags =
+		useSelect( select => select( 'core/editor' ).getEditedPostAttribute( 'tags' ), [] ) || [];
+	const tagObjects = useSelect(
+		select => {
+			return tags
+				.map( tagId => {
+					const tag = select( 'core' ).getEntityRecord( 'taxonomy', 'post_tag', tagId );
+
+					if ( ! tag ) {
+						// Data is not yet loaded
+						loading = true;
+						return;
+					}
+
+					return tag;
+				} )
+				.filter( Boolean ); // Remove undefined values
+		},
+		[ tags ]
+	);
+
+	useEffect( () => {
+		setIsLoadingCategories( loading );
+	}, [ loading ] );
+
+	const postId = useSelect( select => select( 'core/editor' ).getCurrentPostId() );
+	const categoryNames = categoryObjects
+		.filter( cat => cat.id !== 1 )
+		.map( ( { name } ) => name )
+		.join( ', ' );
+	const tagNames = tagObjects.map( ( { name } ) => name ).join( ', ' );
+
+	const contentBefore = useSelect( select => {
+		const editor = select( 'core/block-editor' );
+		const index = editor.getBlockIndex( clientId );
+		return editor.getBlocks().slice( 0, index ) ?? [];
+	} );
+
+	const getSuggestionFromOpenAI = type => {
+		if ( !! content || isLoadingCompletion ) {
+			return;
+		}
+
+		setShowRetry( false );
+		setErrorMessage( false );
+		setIsLoadingCompletion( true );
+
+		const data = {
+			content: createPrompt(
+				currentPostTitle,
+				contentBefore,
+				categoryNames,
+				tagNames,
+				userPrompt,
+				type
+			),
+		};
+
+		tracks.recordEvent( 'jetpack_ai_gpt3_completion', {
+			post_id: postId,
+		} );
+
+		apiFetch( {
+			path: '/wpcom/v2/jetpack-ai/completions',
+			method: 'POST',
+			data: data,
+		} )
+			.then( res => {
+				const result = res.trim();
+				const markdownConverter = new MarkdownIt();
+				setAttributes( { content: result.length ? markdownConverter.render( result ) : '' } );
+				setIsLoadingCompletion( false );
+			} )
+			.catch( e => {
+				if ( e.message ) {
+					setErrorMessage( e.message ); // Message was already translated by the backend
+				} else {
+					setErrorMessage(
+						__(
+							'Whoops, we have encountered an error. AI is like really, really hard and this is an experimental feature. Please try again later.',
+							'jetpack'
+						)
+					);
+				}
+				setShowRetry( true );
+				setIsLoadingCompletion( false );
+			} );
+	};
+	return {
+		getSuggestionFromOpenAI,
+		isLoadingCategories,
+		isLoadingCompletion,
+		setIsLoadingCategories,
+		setShowRetry,
+		showRetry,
+	};
+};
+
+export default useSuggestionsFromOpenAI;
